[level 2] Title: 점 찍기, Time: 20.15 ms, Memory: 79.3 MB -BaekjoonHub

+++ 프로그래머스/2/140107. 점 찍기/점 찍기.java
@@ -1,12 +1,12 @@
 class Solution {
     public long solution(int k, int d) {
         long answer = 0;
-        
-        for (long x = 0; x <= d; x += k) {
+        // 공식 x^2 + y^2 <= d^2
+        // y <= d^2 - x^2 의 제곱근
+        for (int x = 0; x <= d; x += k) {
             double y = Math.sqrt(Math.pow(d, 2) - Math.pow(x, 2));
             answer += (long) (y / k) + 1;
         }
-        
         return answer;
     }
 }
